data/requirements.txt
google-genai
qdrant-client
yt-dlp
opencv-python-headless
beautifulsoup4
requests
python-dotenv
PyYAML
